+++ crates/wolf_engine_events/src/dynamic/mod.rs
@@ -4,6 +4,7 @@ mod event;
 pub use event::*;
 
 pub mod event_loop;
+pub mod events;
 
 mod event_sender;
 pub use event_sender::*;