+++ src/lib.rs
@@ -1,1 +1,2 @@
+#[cfg(feature = "events")]
 pub use wolf_engine_events as events;