+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [0.1]
 
 - Added `dynamic` module.
-- Added `dynamic::Event` trait.
+  - Added `Event` trait.
 - Added `mpsc` module.
   - Added `MpscEventReceiver` struct.
   - Added `MpscEvenSender` struct.