+++ CHANGELOG.md
@@ -21,15 +21,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       - Added `EventReceiver` impl.
       - Added `Default` impl.
       - Added `event_sender()` method.
+  - Added `events` module.
+    - Added `Quit` event.
+    - Added `EventsCleared` event.
   - Added `DynamicEvent` trait.
   - Added `DynamicEventSender` trait.
     - Added auto-impl for `EventSender<DynamicEventBox>` types.
     - Added a provided `send_event()` method.
   - Added `DynamicEventBox` type-def of `Box<dyn DynamicEvent>`.
   - Added `DynamicEvent` derive macro.
-- Added `events` module.
-  - Added `Quit` event.
-  - Added `EventsCleared` event.
 - Added `mpsc` module.
   - Added `MpscEventReceiver` struct.
     - Added `EventReceiver` impl.