+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 
+- Added `dynamic_events` feature.
 - Added `wolf_engine_events` crate.
   - Added public re-export of `wolf_engine_events` as `events`.
 - Added `wolf_engine_codegen` crate.
@@ -15,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [0.1]
 
+- Added `dynamic` feature.
 - Added `dynamic` module.
   - Added `event_loop` module.
     - Added `EventLoop` struct.