+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 
-- Added `dynamic_events` feature.
+- Added `dynamic_events` feature corrisponding to the 
+  `wolf_engine_events/dynamic` feature.
 - Added `wolf_engine_events` crate.
   - Added public re-export of `wolf_engine_events` as `events`.
 - Added `wolf_engine_codegen` crate.
@@ -16,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [0.1]
 
-- Added `dynamic` feature.
+- Added `dynamic` feature to enable / disable the `dynamic` module.
 - Added `dynamic` module.
   - Added `event_loop` module.
     - Added `EventLoop` struct.