+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # [Unreleased]
 
-- Added `dynamic_events` feature corrisponding to the 
-  `wolf_engine_events/dynamic` feature.
+- Added `events` feature to enable / disable the `events` module.
+- Added `dynamic` feature to enable / disable dynamic type features
+  engine-wide.
 - Added `wolf_engine_events` crate.
   - Added public re-export of `wolf_engine_events` as `events`.
 - Added `wolf_engine_codegen` crate.