+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [wolf_engine_events]
 
-### [0.1]
+### [0.1] - 2024-03-07
 
 - Added `dynamic` feature to enable / disable the `dynamic` module.
 - Added `dynamic` module.
@@ -48,11 +48,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [wolf_engine_codegen]
 
-### [0.1]
+### [0.1] - 2024-03-07
 
 - Added `DynamicEvent` derive macro.
 
-# [0.1 - 0.25] 2021-08-11 - 2024-03-02
+# [0.1 - 0.25] - 2021-08-11 - 2024-03-02
 
 These are the old, prototype versions of Wolf Engine, and changes to them
 are mostly undocumented.  You can find these versions in the